requirements.txt
openai
webvtt-py
pysrt
tqdm
python-dotenv